Apply border around book elements
Been a couple months since I have worked on the project so, trying to
find my feet again. Here I create a border around the book elements in
a slightly darker colour to give the books a sort of 3D effect, and make
the elements feel more like books rather than cards on the page.